requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.12.1
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
random_word==1.0.13
reportlab==4.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
